Gateway works with graphql-java-kickstart
https://www.graphql-java-kickstart.com/

+++ graphql-gateway/src/main/kotlin/dev/jmfayard/factsdemo/Fact.kt
@@ -1,3 +1,3 @@
 package dev.jmfayard.factsdemo
 
-data class Fact(val fact: String, val length: Int, val latency: Long?)
+data class Fact(val fact: String, val length: Int, val latency: Int?)
